Update live number of words update in detail VC

+++ Documents/Documents/Model/MGBDocument.m
@@ -20,7 +20,7 @@
     return self;
 }
 
-- (int)wordCount
+- (NSUInteger)wordCount
 {
     return [self.text wordCount];
 }
